Fix no multiple reconnects on close
Changes the default state to CLOSED.

Previously, it was possible to call `connect()` multiple times and open
multiple connections. This was a particular problem if the server ended
the connection. Each time the server ended the connection we'd call n+1
connection events.

This guards against multiple connects by putting a check at the top of
the `connect()` method and ensuring the state is set to CLOSED when
appropriate.